+++ components/layout/default-layout.js
@@ -1,12 +1,12 @@
 import Navbar from './navbar'
-//  import Footer from './footer'
+import Footer from './footer'
 
 export default function DefaultLayout({ children }) {
   return (
     <>
       <Navbar />
       <main>{children}</main>
-      {/* <Footer /> */}
+      <Footer />
     </>
   )
 }