+++ configs/index.js
@@ -1,1 +1,2 @@
-export const dev = true
+export const PORT = 6005
+export const DEV = true