+++ next.config.js
@@ -4,18 +4,15 @@ const nextConfig = {
   images: {
     domains: ['via.placeholder.com', 'localhost'],
   },
+  // avoid cors with proxy
+  async rewrites() {
+    return [
+      {
+        source: '/api/:path*',
+        destination: 'http://localhost:3005/:path*', // Proxy to Backend
+      },
+    ]
+  },
 }
-
-// avoid cors with proxy
-// module.exports = {
-//   async rewrites() {
-//     return [
-//       {
-//         source: '/api/:path*',
-//         destination: 'http://localhost:3005/api/:path*',
-//       },
-//     ]
-//   },
-// }
 
 module.exports = nextConfig