+++ pages/shop/category/[slug].js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function ProductCategory() {
-  return <div>ProductCategory</div>
-}