+++ pages/store/index.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export default function StoreIndex() {
-  return <div>StoreIndex</div>
-}