+++ pages/store/nested.js
@@ -1,3 +1,0 @@
-export default function NestedStore() {
-  return <div>NestedStore</div>
-}