+++ pages/store/nested/index.js
@@ -1,3 +1,0 @@
-export default function NestedStoreIndex() {
-  return <div>NestedStoreIndex</div>
-}